検索画面コントロール化
