+++ StockControlSystem.csproj
@@ -47,11 +47,11 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="frmControlSearch.cs">
+    <Compile Include="CtrFrmSearch.cs">
       <SubType>UserControl</SubType>
     </Compile>
-    <Compile Include="frmControlSearch.Designer.cs">
-      <DependentUpon>frmControlSearch.cs</DependentUpon>
+    <Compile Include="CtrFrmSearch.Designer.cs">
+      <DependentUpon>CtrFrmSearch.cs</DependentUpon>
     </Compile>
     <Compile Include="frmDateTimePicker.cs">
       <SubType>Form</SubType>
@@ -93,8 +93,8 @@
     <Compile Include="pgSelectSQL.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <EmbeddedResource Include="frmControlSearch.resx">
-      <DependentUpon>frmControlSearch.cs</DependentUpon>
+    <EmbeddedResource Include="CtrFrmSearch.resx">
+      <DependentUpon>CtrFrmSearch.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="frmDateTimePicker.resx">
       <DependentUpon>frmDateTimePicker.cs</DependentUpon>
