--- frmControlSearch.cs
@@ -1,20 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-namespace StockControlSystem
-{
-    public partial class frmControlSearch : UserControl
-    {
-        public frmControlSearch()
-        {
-            InitializeComponent();
-        }
-    }
-}
